feat(refactor): Refactor the Scuffle Project (#212)
This commit refactors and cleans up the project getting ready for our
launch in 2025. Prior to this commit, scuffle was going to be a `Live
Streaming Platform` similar to Twitch or YouTube. The project's current
goal is to create an open-source cloud provider similar to Cloudflare,
Mux, AWS or GCP. We want to provide cost effective services that can be
self-hosted easily with minimal configuration, and easy interoperability
between various other providers and our services.