add package to ci
Moves the clippy powerset stuff into the clippy job (only runs on a
brawl run) and then adds a new job that runs `cargo package`. Cargo
Package is a tool ran before `cargo publish` (it runs automatically
unless you do `--no-verify`). This is a sanity check to test if the
crate compiles before uploading it to crates.io. The reason we do this
in a power-set is because we want to make sure that all features of the
crate compile successfully. This is a known issue on crates.io where
sometimes crates break and or dont work depending on the features they
are enabled with.